make $$ more type safe
Previously, an assignment like $$ = 1 would implicitly make the
target an int, even if the intended type was something else, like
a byte. This change creates temp variables of the correct type
as proxy for $$, and eventually assigns them to the %sVAL, which
is the eventual target for $$.

The type safety comes at a cost: if you have any productions that have
actions, then that production must have a specific type. We could
consider tolerating this by using %sVAL without the type.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>